% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{parameter.estimate.table.row}
\alias{parameter.estimate.table.row}
\title{parameter.estimate.table.row}
\usage{
parameter.estimate.table.row(
  name,
  label = NULL,
  units = NULL,
  type = c("Structural", "CovariateEffect", "IIV", "IOV", "RUV", "Unspecified"),
  trans = c("identity", "\%", "exp", "ilogit", "CV\%", "SD (CV\%)"),
  expression = NULL,
  relatedTo = NULL,
  superscript = NULL,
  fixed = NULL,
  value = NULL,
  se = NULL,
  rse = NULL,
  lci95 = NULL,
  uci95 = NULL,
  boot.median = NULL,
  boot.lci = NULL,
  boot.uci = NULL,
  shrinkage = NULL,
  na = "n/a",
  digits = 3,
  indent = TRUE,
  have.bootstrap = !is.null(boot.median),
  columns = c(value = "Estimate", rse = "RSE\%", ci95 = "95\%CI", shrinkage =
    "Shrinkage"),
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
parameter.estimate.table.row
}
