% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{parameter.estimate.table.section}
\alias{parameter.estimate.table.section}
\title{Title}
\usage{
parameter.estimate.table.section(label, ncolumns)
}
\arguments{
\item{ncolumns}{}
}
\value{

}
\description{
Title
}
