% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{parframe}
\alias{parframe}
\title{Create parameter \code{data.frame}}
\usage{
parframe(out, meta, bootstrap = NULL, conf.level = 0.95)
}
\arguments{
\item{out}{paramter data frame from parframe2setup}

\item{meta}{meta information data frame from parframe2setup}

\item{bootstrap}{flag for availability of bootstrap results}

\item{conf.level}{confidence interval of parameter estimates using normal distribution assumptions (in case no bootstrap rasults are available). By default set to 95%.}
}
\description{
Get all relevant parameter information in a data.frame, input the desired transformation for parameters
and finally merge parameter estimates with their corresponding meta information
}
