% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{parframe2setup}
\alias{parframe2setup}
\title{Import files from NONMEM folder to generate paremeter table}
\usage{
parframe2setup(
  run_dir,
  run_prefix,
  runno,
  bootstrap = NULL,
  run_dir.boot = NULL,
  runno.boot = NULL,
  conf.level = 0.95,
  min_suc = TRUE,
  read.boot = NULL,
  boot.obj = NULL,
  yaml.file = NULL,
  yaml.file.name = NULL
)
}
\arguments{
\item{run_dir}{NONMEM model directory}

\item{run_prefix}{NONMEM run number prefix}

\item{runno}{NONMEM run number}

\item{bootstrap}{flag for availability of bootstrap results.}

\item{run_dir.boot}{bootstrap results directory}

\item{runno.boot}{bootstrap NONMEM run number (if changes from runno)}

\item{conf.level}{bootstrap results confidence interval. By default set to 95%.}

\item{read.boot}{flag for reading directly the customized filtering bootstrap raw_results.csv}

\item{boot.obj}{customized filtered bootstrap raw_results.csv data frame passed by user}

\item{yaml.file}{flag for using the yaml.file option. All the non-compulsory fields (ie, label, trans, units, type) will be read from yaml file.}

\item{yaml.file.name}{name of the yaml file.}

\item{min_succ}{filter bootstrap results on minimization successful. By default set to TRUE.}
}
\value{
meta file information (df_m) and parameter information (prm)
}
\description{
reads all relevant inputs (ie, parameter estimates, bootstrap, shrinkage, etc.) using directly model run or with the help of a YAML file.
The parameter name must be provided in the model file. Other fields are optional and can be provided either in (1) the model file or a (2) YAML file.
Few fields needs to be inputted directly from user to complete the table:
  name (compulsory): name of parameter (eg, CL, Vd, nCL, nVd, etc.)
  label:  description of parameter (eg, "Volume of distribution", "Apparent Clearance", etc.)
  units: parameter unit (eg, L/h, 1/hr, etc)
  trans: parameter transformation (ie, %, exp, ilogit, sqrt, CV%)
  type:  Typical Values, Between Subject Variability, Inter Occasion Variability, Residual Error or Covariates (ie respectively, Structural, IIV, IOV, RUV or CovariateEffect)
}
