% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pmxpartab}
\alias{pmxpartab}
\title{Create html table from parframe}
\usage{
pmxpartab(
  parframe,
  meta,
  columns = c(value = "Estimate", rse = "RSE\%", ci95 = "95\%CI", shrinkage =
    "Shrinkage"),
  sections = TRUE,
  section.labels = c(Structural = "Typical Values", CovariateEffect =
    "Covariate Effects", RUV = "Residual Error", IIV = "Between Subject Variability", IOV
    = "Inter-Occasion Variability"),
  show.fixed.to.zero = F,
  na = "n/a",
  digits = 3
)
}
\arguments{
\item{parframe}{object from parframe}

\item{meta}{meta information data frame from parframe2setup}

\item{columns}{which information to display in the table}

\item{sections}{flag for having separated sections displayed}

\item{section.labels}{which sections will be displayed}

\item{show.fixed.to.zero}{flag to show or not parameter fixed to zero. By default will be not displayed.}

\item{na}{how NA will be displayed}

\item{digits}{how many digits will be displayed}
}
\description{
Uses the output from parframe to generate a parameter estimates table in HTML
}
