% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{parse_parameter_description}
\alias{parse_parameter_description}
\title{Parse parameter description}
\usage{
parse_parameter_description(string)
}
\arguments{
\item{string}{Character string}
}
\value{
Character; parameter description
}
\description{
Parse parameter description
}
